+++ src/test/java/camp/nextstep/edu/kitchenpos/menugroup/dao/InMemoryMenuGroupDao.java
@@ -1,7 +1,7 @@
-package camp.nextstep.edu.kitchenpos.bo.mock;
+package camp.nextstep.edu.kitchenpos.menugroup.dao;
 
-import camp.nextstep.edu.kitchenpos.dao.MenuGroupDao;
-import camp.nextstep.edu.kitchenpos.model.MenuGroup;
+import camp.nextstep.edu.kitchenpos.menugroup.dao.MenuGroupDao;
+import camp.nextstep.edu.kitchenpos.menugroup.model.MenuGroup;
 
 import java.util.*;
